backend/pyproject.toml
[tool.poetry]
name = "Spring"
version = "0.1.0"
description = "A learning resource manager"
authors = ["Deborah Idumu <deborahidumu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
pydantic = "^2.10.6"
asyncpg = "^0.30.0"
python-multipart = "^0.0.20"
pyjwt = "^2.10.1"
passlib = "1.7.4"
python-dotenv = "^1.0.1"
bcrypt = "4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
